refactor: types 나누기
