+++ frontend/src/hooks/fetch/useFetchProductData.ts
@@ -1,6 +1,6 @@
 import { PRODUCT_SUMMARY_URL } from '@/api/api.ts';
 import useFetchData from '@/hooks/fetch/useFetchData.ts';
-import type { SelectedItemInfo } from '@/types/allTypes.ts';
+import type { SelectedItemInfo } from '@/types/order/types.ts';
 
 export default function useFetchProductData(id: number) {
   const { data, isLoading, error } = useFetchData<SelectedItemInfo>(
