+++ frontend/src/hooks/fetch/useFetchRanking.ts
@@ -1,6 +1,6 @@
 import { RANKING_URL } from '@/api/api.ts';
 import useFetchData from '@/hooks/fetch/useFetchData.ts';
-import type { Product } from '@/types/allTypes.ts';
+import type { Product } from '@/types/themes/types.ts';
 
 export default function useFetchRanking(targetType: string, rankType: string) {
   const { data } = useFetchData<Product>(['ranking', targetType, rankType], RANKING_URL, {
