+++ frontend/src/hooks/fetch/useFetchThemesInfo.ts
@@ -1,6 +1,6 @@
 import useFetchData from '@/hooks/fetch/useFetchData.ts';
 import { THEMES_INFO } from '@/api/api.ts';
-import type { ThemeInfo, UseFetchThemesInfoResult } from '@/types/allTypes.ts';
+import type { ThemeInfo, UseFetchThemesInfoResult } from '@/types/themes/types.ts';
 
 export default function useFetchThemesInfo(themesId: number): UseFetchThemesInfoResult {
   const url = THEMES_INFO(themesId);
