+++ frontend/src/styles/theme.ts
@@ -1,7 +1,7 @@
-import { colors } from '@/types/theme/palette';
-import { typography } from '@/types/theme/typography';
-import { spacing } from '@/types/theme/spacing';
-import { semantic } from '@/types/theme/semantic.ts';
+import { colors } from '@/types/style/palette';
+import { typography } from '@/types/style/typography';
+import { spacing } from '@/types/style/spacing';
+import { semantic } from '@/types/style/semantic.ts';
 
 export const theme = {
   colors,
