+++ frontend/src/types/order/types.ts
@@ -1,0 +1,7 @@
+export interface SelectedItemInfo {
+  id: string | number;
+  name: string;
+  brandName: string;
+  price: number;
+  imageURL: string;
+}
