+++ frontend/src/types/style/palette.ts
